lock gltf
