+++ src/editor/components/scenegraph/Toolbar.js
@@ -13,7 +13,6 @@ import {
   Edit24Icon
 } from '../../icons';
 import Events from '../../lib/Events';
-import { saveBlob } from '../../lib/utils';
 import { Button, ProfileButton } from '../components';
 import { uploadThumbnailImage } from '../modals/ScreenshotModal/ScreenshotModal.component.jsx';
 import { sendMetric } from '../../services/ga.js';
@@ -21,34 +20,6 @@ import posthog from 'posthog-js';
 import { UndoRedo } from '../components/UndoRedo';
 import debounce from 'lodash-es/debounce';
 // const LOCALSTORAGE_MOCAP_UI = "aframeinspectormocapuienabled";
-
-function filterHelpers(scene, visible) {
-  scene.traverse((o) => {
-    if (o.userData.source === 'INSPECTOR') {
-      o.visible = visible;
-    }
-  });
-}
-
-function getSceneName(scene) {
-  return scene.id || slugify(window.location.host + window.location.pathname);
-}
-
-/**
- * Slugify the string removing non-word chars and spaces
- * @param  {string} text String to slugify
- * @return {string}      Slugified string
- */
-function slugify(text) {
-  return text
-    .toString()
-    .toLowerCase()
-    .replace(/\s+/g, '-') // Replace spaces with -
-    .replace(/[^\w-]+/g, '-') // Replace all non-word chars with -
-    .replace(/--+/g, '-') // Replace multiple - with single -
-    .replace(/^-+/, '') // Trim - from start of text
-    .replace(/-+$/, ''); // Trim - from end of text
-}
 
 /**
  * Tools and actions.
@@ -351,37 +322,6 @@ export default class Toolbar extends Component {
   // openViewMode() {
   //   AFRAME.INSPECTOR.close();
   // }
-
-  static exportSceneToGLTF() {
-    try {
-      sendMetric('SceneGraph', 'exportGLTF');
-      const sceneName = getSceneName(AFRAME.scenes[0]);
-      const scene = AFRAME.scenes[0].object3D;
-      posthog.capture('export_scene_to_gltf_clicked', {
-        scene_id: STREET.utils.getCurrentSceneId()
-      });
-
-      filterHelpers(scene, false);
-      AFRAME.INSPECTOR.exporters.gltf.parse(
-        scene,
-        function (buffer) {
-          filterHelpers(scene, true);
-          const blob = new Blob([buffer], { type: 'application/octet-stream' });
-          saveBlob(blob, sceneName + '.glb');
-        },
-        function (error) {
-          console.error(error);
-        },
-        { binary: true }
-      );
-      STREET.notify.successMessage('3DStreet scene exported as glTF file.');
-    } catch (error) {
-      STREET.notify.errorMessage(
-        `Error while trying to save glTF file. Error: ${error}`
-      );
-      console.error(error);
-    }
-  }
 
   toggleScenePlaying = () => {
     if (this.state.isPlaying) {
